+++ src/javascript/controls/notifications.js
@@ -13,31 +13,59 @@ License > https://github.com/Graviton-Code-Editor/Graviton-App/blob/master/LICEN
 
 /*
 
+SIMPLE:
+
 new Notification('Title','This is a content')
+
+
+WITH BUTTONS:
+
+new Notification("Notification","This notification has buttons",
+  {
+    "Install":function(){
+      console.log("Installing?")
+    },
+    "Cancel":function(){
+      closeNotification(this.parentElement)
+    }
+  }
+)
 
 */
 module.exports = {
-  Notification: function(title, message) {
+  Notification: function(title, message,buttons) {
     if (_notifications.length >= 3) {//Remove one notification in case there are 3
       _notifications[0].remove();
       _notifications.splice(0, 1);
     }
-    const textID = Math.random();
     const body = document.createElement("div");
     body.classList.add("notificationBody");
     body.setAttribute("id", _notifications.length);
     body.innerHTML = `
-      <button  onclick="closeNotification(this)">
+      <button class=close onclick="closeNotification(this)">
           ${icons["close"]}
       </button>
       <h1>${title}</h1>
       <div>
-          <p id="notification_message${textID}"></p>
-      </div>`;
+          <p >${message}</p>
+      </div>
+      ${buttons!=undefined?"<span class=line_space_menus></span>":""}
+      <div>
+          
+      </div> `;
+      if(buttons!=undefined){
+        Object.keys(buttons).map(function(key){
+          const id = Math.random();
+          const button = document.createElement("button");
+          button.innerText = key;
+          sleeping(1).then(() => {
+            button.addEventListener("click",buttons[key]) 
+            button.setAttribute("onClick","closeNotification(this.parentElement)")            
+          });
+          body.children[4].appendChild(button)
+        })
+      }
     document.getElementById("notifications").appendChild(body);
-    document.getElementById(
-      `notification_message${textID}`
-    ).innerText = message;
     _notifications.push(body);
     const wait = setTimeout(() => {
       for (i = 0; i < _notifications.length; i++) {
@@ -46,7 +74,7 @@ module.exports = {
           body.remove();
         }
       }
-    }, 7000); //Wait 7 seconds until the notification automatically deletes it self
+    }, 50000); //Wait 7 seconds until the notification automatically deletes it self
   },
   closeNotification: function(element) {
     for (i = 0; i < _notifications.length; i++) {
@@ -57,3 +85,5 @@ module.exports = {
     }
   }
 };
+
+